backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.13
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.8.1.78
# tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
mediapipe==0.10.11
openai==1.3.0
python-dotenv==1.0.0 
